Throw when querying non-scalar objects without a selection set.
This would have helped diagnose the cause of #4025, which is that query
fields with non-scalar types (or list-of-non-scalar types) must have a
selection set.

In other words, given a schema like

  type Query {
    selectedMessageList: [SelectedMessage]
  }

a query like

  query getSelectedMessageList {
    selectedMessageList @client
  }

is invalid since it doesn't specify which fields of the SelectedMessage
objects should be included in the list.